refactor structure to use repo index file
