+++ src/repos/common/index.js
@@ -1,11 +1,9 @@
 import getById from './get-by-id.js'
 import getByProperty from './get-by-property.js'
-import getByReference from '../comms/get-by-reference.js'
 import saveEvent from './save-event.js'
 
 export {
   getById,
   getByProperty,
-  saveEvent,
-  getByReference
+  saveEvent
 }
