+++ test/integration/narrow/repos/common/get-by-id.test.js
@@ -1,5 +1,5 @@
 import { describe, test, expect, beforeEach } from 'vitest'
-import getById from '../../../../../src/repos/common/get-by-id.js'
+import { getById } from '../../../../../src/repos/common/index.js'
 import db from '../../../../../src/data/db.js'
 import v1CommsMessage from '../../../../mocks/comms-message/v1.js'
 import { NotFoundError } from '../../../../../src/errors/not-found-error.js'
